frontend/requirements.txt
streamlit
pandas
requests
python-dotenv
google-cloud-bigquery
